+++ src/modules/auth/use-cases/register.use-case.ts
@@ -3,13 +3,10 @@ import * as bcrypt from 'bcrypt';
 import type { Request } from 'express';
 import { UserRole } from 'generated/prisma/enums';
 import { EnumAuthError } from '../consts/auth.errors';
-import { AuthSessionRequestDto } from '../dto/auth-session/auth-session.request.dto';
+import { AuthService } from '../services/auth.service';
 import { RegisterDto } from '../dto/register/register-request.dto';
 import { RegisterResponseDto } from '../dto/register/register-response.dto';
-import { AuthContextService } from '../services/auth-context.service';
-import { AuthRequestMetaService } from '../services/auth-request-meta.service';
-import { BCRYPT_SALT_ROUNDS, TokenService } from '@/modules/token';
-import { RoleContextService } from '@/modules/role-context';
+import { BCRYPT_SALT_ROUNDS } from '@/modules/token';
 import { UserService } from '@/modules/user';
 
 /**
@@ -20,10 +17,7 @@ import { UserService } from '@/modules/user';
 export class RegisterUseCase {
 	constructor(
 		private readonly userService: UserService,
-		private readonly roleContextService: RoleContextService,
-		private readonly authContextService: AuthContextService,
-		private readonly authRequestMetaService: AuthRequestMetaService,
-		private readonly tokenService: TokenService,
+		private readonly authService: AuthService,
 	) {}
 
 	private assertRegisterableRole(role: UserRole): asserts role is typeof UserRole.CANDIDATE | typeof UserRole.EMPLOYER {
@@ -50,29 +44,11 @@ export class RegisterUseCase {
 
 		/** Пока без письма: сразу «подтверждённая» учётка. */
 		const activated = await this.userService.update(created.id, { isActivated: true });
-
-		const roleContext = await this.roleContextService.findFirstForUserWithHrRole(activated.id, dto.role);
-		if (!roleContext) {
-			throw new UnauthorizedException(EnumAuthError.ROLE_NOT_FOUND);
-		}
-
-		const buildData: AuthSessionRequestDto = {
-			roleContextId: roleContext.id,
-			userId: activated.id,
-			userRole: roleContext.userRole,
-		};
-		const payload = await this.authContextService.buildAccessPayload(buildData);
-		const requestMeta = this.authRequestMetaService.fromRequest(request);
-		const pair = await this.tokenService.generateTokenPair(payload, requestMeta);
-
+		const pair = await this.authService.getAuthenticatedToken(request, activated.id, dto.role);
+		const user = this.authService.getAuthenticatedUser(activated, dto.role);
 		return {
 			tokens: pair,
-			user: {
-				id: activated.id,
-				email: activated.email,
-				role: roleContext.userRole,
-				isActivated: activated.isActivated,
-			},
+			user: user,
 		};
 	}
 }
